+++ NEWS.md
@@ -1,7 +1,9 @@
 # conductor 0.1.0.9000
 
 * `conductor` now scrolls smoothly between elements (#19). 
-* Update `shepherd.js` to v.10.0.0
+* Update `shepherd.js` to v.10.0.1
+* Method `isCentered()` is no longer supported because it was removed in 
+  `shepherd.js` v.10.0.0.
 
 # conductor 0.1.0